\name{rheologyEvaluation-package}
\alias{rheologyEvaluation-package}
\alias{plot.counts}
\docType{package}
\title{
rheologyEvaluation
}
\description{
Some utility functions for reading and analyzing rheology files from Rheowin
}
\details{
\tabular{ll}{
Package: \tab rheologyEvaluation\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2018-10-17\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
Thomas Braschler

Maintainer: thomas.braschler@gmail.com
}
\keyword{ package }


